% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-series.R
\name{add_all_layers}
\alias{add_all_layers}
\title{Add extra a list of single argument plotly layer functions}
\usage{
add_all_layers(plt, layers)
}
\arguments{
\item{plt:}{plotly: existing plot to add layers to}

\item{layers:}{list of functions: list of single argument functions}
}
\value{
plotly plot
}
\description{
All functions should take a single plot argument
}
