% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily-cases-manipulation.R
\name{add_ma_smoothing}
\alias{add_ma_smoothing}
\title{Takes daily cases time series and performaing moving average smoothing}
\usage{
add_ma_smoothing(
  daily_cases_ts,
  cols_used = c("NEW_CASES", "NEW_DEATHS"),
  orders = c(3, 7, 30, 60)
)
}
\arguments{
\item{daily_cases_ts}{tibble: daily cases time series (1 date per row)}

\item{orders}{int vector: order of smoothing (each element will produce a new column)}

\item{cols_used:}{character vector: columns to be smoothed}
}
\value{
A tibble of the daily cases time series smoothed with original columns
}
\description{
Takes daily cases time series and performaing moving average smoothing
}
\examples{
\dontrun{
daily_cases <- get_daily_cases()
daily_cases_ts <- get_daily_agg(daily_cases, cols_sel = c("NEW_CASES", "NEW_DEATHS"))
daily_cases_ts_smoothed <- add_ma_smoothing(daily_cases_ts)

daily_cases <- get_daily_cases()
daily_cases_ts <- get_daily_agg(daily_cases, cols_sel = c("NEW_CASES"))
daily_cases_ts_smoothed <- add_ma_smoothing(daily_cases_ts, cols_used = c("NEW_CASES"), orders = c(10, 20, 100))
plot(daily_cases_ts_smoothed$NEW_CASES, type = 'l')
lines(daily_cases_ts_smoothed$NEW_CASES_MA_10DAY, col = "red")
lines(daily_cases_ts_smoothed$NEW_CASES_MA_20DAY, col = "green")
lines(daily_cases_ts_smoothed$NEW_CASES_MA_100DAY, col = "purple")
}

}
