% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-series.R
\name{blank_ts_plot}
\alias{blank_ts_plot}
\title{Generates a blank canvas time series plot}
\usage{
blank_ts_plot(data = data.frame(), labels = list(), colour_palette = list())
}
\arguments{
\item{data}{data frame: used when adding different data to an existing plot}

\item{labels:}{list: e.g. \code{list(xaxis = "Date", yaxis = "Magnitude", title = "New Plot")}}

\item{colour_palette:}{list: allows custom color pallet that overrides global
one. (see config/colour-palette.yaml for changable colours)}
}
\value{
plotly plot_ly config layout
}
\description{
This will need to have data added to it
}
\examples{
\dontrun{
blank_ts_plot()
}
}
