% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dashboard-body.R
\name{body_server}
\alias{body_server}
\title{Dashboard body server function}
\usage{
body_server(id, daily_cases, vax, population)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}

\item{daily_cases}{tibble: daily cases from WHO}

\item{vax}{TODO}

\item{population}{TODO}
}
\description{
Dashboard body server function
}
