% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dashboard-body.R
\name{body_ui}
\alias{body_ui}
\title{Dashboard body ui function}
\usage{
body_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Dashboard body ui function
}
