% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-dataset-daily-cases.R
\name{build_daily_cases_dataset}
\alias{build_daily_cases_dataset}
\title{Download and write the daily cases dataset}
\usage{
build_daily_cases_dataset()
}
\description{
Download and write the daily cases dataset
}
