% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-time-series.R
\name{build_ts_trace_fn}
\alias{build_ts_trace_fn}
\title{Higher order function that builds a configured \code{plotly::add_trace}.}
\usage{
build_ts_trace_fn(x, data = NULL, type = "scatter", mode = "lines")
}
\arguments{
\item{x}{string: the x variable. A string is converted to a formula like ~X_AXIS}

\item{data}{data frame: used when adding different data to an existing plot}

\item{type}{string: type arg passed to trace type}

\item{mode}{string: mode arg passed to trace type}
}
\value{
function: configured add_trace
}
\description{
Useful for consistent trace adding.
}
\examples{
\dontrun{
add_ts_trace <- build_ts_trace_fn(x = "DATE")
}
}
