% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-dataset-vaccination.R
\name{build_vaccination_dataset}
\alias{build_vaccination_dataset}
\title{Download and write vaccination dataset}
\usage{
build_vaccination_dataset()
}
\description{
Download and write vaccination dataset
}
