% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headline_figures.R
\name{calc_cases}
\alias{calc_cases}
\title{Calculates the latest new case numbers}
\usage{
calc_cases(daily_cases)
}
\arguments{
\item{daily_cases}{tibble: daily cases from WHO}
}
\value{
list: new_cases = latest number of new cases, total_cases: total confirmed
cases, perc_inc = daily \% increase
}
\description{
Calculates the latest new case numbers
}
\examples{
\dontrun{
daily_cases <- get_daily_cases()
output <- calc_cases(daily_cases)
}

}
\seealso{
\code{calc_new_and_total(daily_cases, new_col, total_col)}
}
