% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headline-figures.R
\name{calc_deaths}
\alias{calc_deaths}
\title{Calculates the latest new death numbers}
\usage{
calc_deaths(daily_cases)
}
\arguments{
\item{daily_cases}{tibble: daily cases from WHO}
}
\value{
list: new = latest number of new deaths, total: total confirmed
deaths, perc_inc = daily \% increase
}
\description{
Calculates the latest new death numbers
}
\examples{
\dontrun{
daily_cases <- get_daily_cases()
output <- calc_deaths(daily_cases)
}

}
\seealso{
\code{calc_new_and_total(daily_cases, new_col, total_col)}
}
