% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headline_figures.R
\name{calc_new_and_total}
\alias{calc_new_and_total}
\title{Calculates the latest new case or death numbers}
\usage{
calc_new_and_total(daily_cases, new_col, total_col)
}
\arguments{
\item{daily_cases}{tibble: daily cases from WHO}
}
\value{
list: new = latest number of new cases, total: total confirmed
cases, perc_inc = daily \% increase
}
\description{
Note that this is the latest new case numbers for each country. Lets say you
have Australia, Brazil and Chile. If Australia and Brazil are up-to-date and
have info for the Friday, but Chile's latest info is for the previous
Wednesday, the new case numbers will include Australia and Brazil's Friday
numbers as well as Brazil's Wednesday numbers.
}
\details{
This could cause new cases to "persist" until later information comes in, but
it prevents a sudden drop in case numbers due to lack of reporting. Hence the
reason to include the latest numbers instead of excluding that country.
}
\examples{
\dontrun{
daily_cases <- get_daily_cases()
deaths <- calc_new_and_total(  daily_cases = daily_cases
                             , new_col = "NEW_DEATHS"
                             , total_col = "CUMULATIVE_DEATHS")
}

}
