% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak-rating.R
\name{calculate_outbreak_rating}
\alias{calculate_outbreak_rating}
\title{Calculates the outbreak rating based on new cases to population ratio.}
\usage{
calculate_outbreak_rating(new_cases, population)
}
\arguments{
\item{new_cases}{numeric: number of newly reported cases}

\item{population:}{numeric: population at the time of newly reported cases}
}
\value{
integer between 0 and 100
}
\description{
Calculates the outbreak rating based on new cases to population ratio.
}
\examples{
\dontrun{
outbreak_rating <- calculate_outbreak_rating(new_cases = 25, population = 100000)
}
}
\seealso{
\code{generate_daily_outbreak_ratings}, \code{describe_outbreak}
}
