% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common-docs.R
\docType{data}
\name{common_docs}
\alias{common_docs}
\title{Common documentation}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
common_docs
}
\arguments{
\item{daily_cases}{dataframe: daily cases from WHO}

\item{vax}{dataframe}

\item{population}{dataframe}

\item{daily_cases_filt}{reactive dataframe: daily cases dataset}

\item{vax_filt}{reactive dataframe}

\item{population_filt}{reactive dataframe: world population dataset}

\item{latest_rows_by_country}{reactive dataframe}

\item{latest_outbreak_rating_by_country}{reactive dataframe}
}
\description{
Common documentation
}
\keyword{internal}
