% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dashboard-controlbar.R
\name{controlbar_sever}
\alias{controlbar_sever}
\title{Dashboard controlbar server function}
\usage{
controlbar_sever(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Dashboard controlbar server function
}
