% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dashboard-controlbar.R
\name{controlbar_ui}
\alias{controlbar_ui}
\title{Dashboard controlbar ui function}
\usage{
controlbar_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Dashboard controlbar ui function
}
