% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-country-comp.R
\name{country_comp_server}
\alias{country_comp_server}
\title{Page: Country comparison server function}
\usage{
country_comp_server(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Country comparison server function
}
