% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-country-comp.R
\name{country_comp_ui}
\alias{country_comp_ui}
\title{Page: Country comparison ui function}
\usage{
country_comp_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Country comparison ui function
}
