% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-data-explorer.R
\name{data_explorer_server}
\alias{data_explorer_server}
\title{Page: Data exploration server function}
\usage{
data_explorer_server(id, daily_cases, vax, population)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}

\item{daily_cases}{tibble: daily cases from WHO}

\item{vax}{TODO}

\item{population}{TODO}
}
\description{
Page: Data exploration server function
}
