% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-data-explorer.R
\name{data_explorer_ui}
\alias{data_explorer_ui}
\title{Page: Data exploration ui function}
\usage{
data_explorer_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Data exploration ui function
}
