% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-data-sources.R
\name{data_sources_server}
\alias{data_sources_server}
\title{Page: Data sources server function}
\usage{
data_sources_server(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Data sources server function
}
