% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-data-sources.R
\name{data_sources_ui}
\alias{data_sources_ui}
\title{Page: Data sources ui function}
\usage{
data_sources_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Data sources ui function
}
