% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak-rating.R
\name{describe_outbreak}
\alias{describe_outbreak}
\title{Generates a description based on the outbreak rating.}
\usage{
describe_outbreak(outbreak_rating)
}
\arguments{
\item{outbreak_rating}{integer: outbreak rating generated by \code{calculate_outbreak_rating}}
}
\value{
string describing the outbreak rating
}
\description{
Based on thresholds.
}
\examples{
\dontrun{
describe_outbreak(c(0, 1, 29, 100, -1))
}
}
\seealso{
\code{generate_daily_outbreak_ratings}, \code{calculate_outbreak_rating}
}
