% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-formatting.R
\name{format_num}
\alias{format_num}
\title{Formats a number to more readable}
\usage{
format_num(num, digits = NULL, short = FALSE)
}
\arguments{
\item{num}{integer or numeric scalar: number will be formatted to string}

\item{digits}{a positive integer indicating how many significant digits are to be used for numeric and complex x. The default, NULL, uses getOption("digits").}

\item{short}{if TRUE trims the string down and applies a suffix. E.g. 11100000 -> 11.1M}
}
\value{
string: string of a number formatted nicely
}
\description{
Formats a number to more readable
}
\examples{
\dontrun{
format_num(1111, digits = 1, short = TRUE)
format_num(123456789, short = FALSE)
format_num(1234567890, digits = 3, short = TRUE)
}

}
