% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh-data.R
\name{generate_backup_path}
\alias{generate_backup_path}
\title{Updates a path to have timestamp prefix in front of the file name.}
\usage{
generate_backup_path(path)
}
\arguments{
\item{path}{string: a string representing a file path}
}
\value{
a string of a file path
}
\description{
Updates a path to have timestamp prefix in front of the file name.
}
