% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_rating.R
\name{generate_daily_outbreak_ratings}
\alias{generate_daily_outbreak_ratings}
\title{Generates a dataset including daily outbreak ratings.}
\usage{
generate_daily_outbreak_ratings(daily_cases_w_pop)
}
\arguments{
\item{daily_cases}{data frame: includes DATE_REPOTED, COUNTRY_CODE and NEW_CASES columns}
}
\value{
tibble dataset with date, country, outbreak rating and outbreak description
}
\description{
Generates a dataset including daily outbreak ratings.
}
\examples{
\dontrun{
daily_cases_sel <- get_daily_cases() \%>\% dplyr::select(DATE_REPORTED, COUNTRY_CODE, NEW_CASES)
population_sel <- get_population() \%>\% dplyr::select(DATE_REPORTED, COUNTRY_CODE, COUNTRY, POPULATION)
daily_cases_w_pop <- dplyr::left_join(daily_cases_sel, population_sel, by = c("DATE_REPORTED", "COUNTRY_CODE"))
outbreak_ratings = generate_daily_outbreak_ratings(daily_cases_w_pop)
}
}
\seealso{
\code{calculate_outbreak_rating}, \code{describe_outbreak}
}
