% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-world-map.R
\name{generate_map_labels}
\alias{generate_map_labels}
\title{Generates html labels for world map}
\usage{
generate_map_labels(dat)
}
\arguments{
\item{dat}{dataframe with COUNTRY, OUTBREAK_RATING and OUTBREAK_DESC columns}
}
\description{
Generates html labels for world map
}
