% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak-rating.R
\name{generate_outbreak_rating_shading_layers}
\alias{generate_outbreak_rating_shading_layers}
\title{Generates the shading layers for the outbreak rating plot}
\usage{
generate_outbreak_rating_shading_layers(xmin, xmax)
}
\arguments{
\item{xmin}{date as a string in "yyyy-mm-dd": where the plot starts on the x-axis}

\item{xmax}{date as a string in "yyyy-mm-dd": where the plot ends on the x-axis}
}
\value{
list: list of layers for plotly plot
}
\description{
Generates the shading layers for the outbreak rating plot
}
\examples{
\dontrun{
generate_outbreak_rating_shading_layers("2022-03-01", "2022-06-10")
}

}
\seealso{
\code{get_outbreak_rating_types}
}
