% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-dataset-population.R
\name{generate_population_data}
\alias{generate_population_data}
\title{Generate the population dataset}
\usage{
generate_population_data(
  daily_cases = get_daily_cases(),
  vacc = get_vaccination()
)
}
\description{
This depends on daily cases and population tasks being complete
}
\details{
NOTE: this feature will not be generated in the near future. It will
instead be retrieved from a reliable data source (such as OECD).
In the meantime will we assume the population is static over time (which
is a very rough assumption).
Steps for this are
\enumerate{
\item Estimate the current population from vaccination data (since it has per 100 people figures)
\item Get historical dates from daily cases
\item Assume the current population has not changed has been static for all dates
\item For any countries that cannot have their population calculated, estimate from their total deaths
}
}
