% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily-cases-manipulation.R
\name{get_daily_agg}
\alias{get_daily_agg}
\title{Takes daily cases and returns the daily aggregation}
\usage{
get_daily_agg(
  daily_cases,
  agg = sum,
  cols_sel = c("NEW_CASES", "NEW_DEATHS"),
  days_back = 365 * 2
)
}
\arguments{
\item{daily_cases}{tibble: daily cases from WHO}

\item{agg}{function: aggregation function. e.g. sum or mean.}

\item{days_back}{int: number of days back the series should extent to}

\item{cols_sel:}{character vector: aggregation columns}
}
\value{
A tibble of the daily cases aggregated
}
\description{
Takes daily cases and returns the daily aggregation
}
\examples{
\dontrun{
daily_cases <- get_daily_cases()
get_daily_agg(daily_cases, agg = sum, cols_sel = c("NEW_CASES"), days_back = 30)
get_daily_agg(daily_cases, agg = mean)
}
}
