% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data-from-storage.R
\name{get_daily_cases_path}
\alias{get_daily_cases_path}
\title{Get storage location for daily cases dataset}
\usage{
get_daily_cases_path()
}
\description{
Get storage location for daily cases dataset
}
