% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh-data.R
\name{get_data_source_config}
\alias{get_data_source_config}
\title{Source and return the data source config object}
\usage{
get_data_source_config()
}
\value{
A list with data source configurations
}
\description{
Source and return the data source config object
}
\examples{
get_data_source_config()

}
