% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_cases_manipulation.R
\name{get_latest_info_by_country}
\alias{get_latest_info_by_country}
\title{Takes daily cases and returns the latest rows for each country}
\usage{
get_latest_info_by_country(daily_cases, lag = 0)
}
\arguments{
\item{daily_cases}{tibble: daily cases from WHO}

\item{lag}{int: number of days back the latest info should be lagged. I.e. 1
day back will give the latest info as if running yesterday. Can be positive
or negative, but will always go back.}
}
\value{
A tibble of the daily cases with the latest dates
}
\description{
Takes daily cases and returns the latest rows for each country
}
\examples{
\dontrun{
daily_cases <- get_daily_cases()
latest_rows_by_country <- get_latest_info_by_country(daily_cases, lag = 1)
}
}
