% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak_rating.R
\name{get_outbreak_rating_types}
\alias{get_outbreak_rating_types}
\title{Generates the labels, descriptions and thresholds for each outbreak rating type.}
\usage{
get_outbreak_rating_types(label = NULL, color = NULL)
}
\arguments{
\item{label}{string: filters the color and threshold based on the associated label}

\item{color}{string: filters the label and threshold based on the associated color}
}
\value{
list of thresholds, labels and colors
}
\description{
Generates the labels, descriptions and thresholds for each outbreak rating type.
}
\examples{
\dontrun{
get_outbreak_rating_types()
get_outbreak_rating_types(label = "Low")
get_outbreak_rating_types(color = "maroon")
get_outbreak_rating_types("Maximum")
}
}
\seealso{
\code{describe_outbreak}
}
