% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data-from-storage.R
\name{get_population_path}
\alias{get_population_path}
\title{Get storage location for population dataset}
\usage{
get_population_path()
}
\description{
Get storage location for population dataset
}
