% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data-from-storage.R
\name{get_vaccination_path}
\alias{get_vaccination_path}
\title{Get storage location for vaccination dataset}
\usage{
get_vaccination_path()
}
\description{
Get storage location for vaccination dataset
}
