% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dashboard-header.R
\name{header_server}
\alias{header_server}
\title{Dashboard header server function}
\usage{
header_server(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Dashboard header server function
}
