% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dashboard-header.R
\name{header_ui}
\alias{header_ui}
\title{Dashboard header ui function}
\usage{
header_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Dashboard header ui function
}
