% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headline-value-box.R
\name{headline_value_box}
\alias{headline_value_box}
\title{Create a headline value box}
\usage{
headline_value_box(
  subtitle,
  new_cases,
  total_cases,
  perc_inc,
  icon,
  color = "light-blue"
)
}
\arguments{
\item{subtitle}{string: subtitle of value box e.g. "New Cases"}

\item{new_cases}{numeric: increase value}

\item{total_cases}{numeric: total value}

\item{perc_inc}{numeric: percentage increase}

\item{icon}{icon tag}
}
\description{
Create a headline value box
}
