% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-formatting.R
\name{inc_or_dec}
\alias{inc_or_dec}
\title{Checks if a number is positive or negative and returns increase or decrease}
\usage{
inc_or_dec(num)
}
\arguments{
\item{num}{integer or numeric scalar: number will be formatted to string}
}
\value{
string: either increase, decrease or ""
}
\description{
Checks if a number is positive or negative and returns increase or decrease
}
\examples{
\dontrun{
inc_or_dec(1)
nc_or_dec(-1)
inc_or_dec(0)
}

}
