% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_base_utilties.R
\name{make_formula}
\alias{make_formula}
\title{Makes a formula object from a string}
\usage{
make_formula(x, prefix = "", suffix = "")
}
\arguments{
\item{x}{string: any string, usually a column name.}

\item{prefix}{string: prefix of the formula}

\item{suffix}{string: suffix of the formula}
}
\value{
a formula. e.g. ~DATE
}
\description{
Makes a formula object from a string
}
\examples{
\dontrun{
make_formula("DATE", "~")
}

}
