% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreak-rating.R
\name{make_outbreak_rating_shading_layer_fn}
\alias{make_outbreak_rating_shading_layer_fn}
\title{Generates a layout function to write the layers of the background for the
outbreak rating plotly plot}
\usage{
make_outbreak_rating_shading_layer_fn(xmin, xmax)
}
\arguments{
\item{xmin}{date as a string in "yyyy-mm-dd": where the plot starts on the x-axis}

\item{xmax}{date as a string in "yyyy-mm-dd": where the plot ends on the x-axis}
}
\value{
function: configured layout function with colored backgrounds
}
\description{
A higher order function that generates a configured layout function for use
with plotly. This is very bespoke to the outbreak rating plot and should not
be used in other plots.
}
\examples{
\dontrun{
time_series_plot_server(
  id = "plot_ts_outbreak",
  ts_data = daily_outbreak_rating_ts,
  labels = list(yaxis = "Outbreak Rating", box_title = "Outbreak Rating", xaxis = ""),
  deselected_traces = c("OUTBREAK_RATING", "OUTBREAK_RATING_MA_30DAY", "OUTBREAK_RATING_MA_90DAY", "OUTBREAK_RATING_MA_180DAY"),
  make_outbreak_rating_shading_layer_fn( # produces the layers for background color shading
    xmin = min(daily_outbreak_rating_ts()[["DATE_REPORTED"]]),
    xmax = max(daily_outbreak_rating_ts()[["DATE_REPORTED"]])
  )
)
}
}
\seealso{
\code{get_outbreak_rating_types}, \code{generate_outbreak_rating_shading_layers}
}
