% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-metrics.R
\name{metrics_server}
\alias{metrics_server}
\title{Page: Metrics server function}
\usage{
metrics_server(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Metrics server function
}
