% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-metrics.R
\name{metrics_ui}
\alias{metrics_ui}
\title{Page: Metrics ui function}
\usage{
metrics_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Metrics ui function
}
