% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common-docs.R
\docType{data}
\name{module_docs}
\alias{module_docs}
\title{Module documentation}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
module_docs
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Module documentation
}
\keyword{internal}
