% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-news-feed.R
\name{news_feed_server}
\alias{news_feed_server}
\title{Page: News feed server function}
\usage{
news_feed_server(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: News feed server function
}
