% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-news-feed.R
\name{news_feed_ui}
\alias{news_feed_ui}
\title{Page: News feed ui function}
\usage{
news_feed_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: News feed ui function
}
