% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-overview.R
\name{overview_ui}
\alias{overview_ui}
\title{Page: Overview ui function}
\usage{
overview_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Overview ui function
}
