% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rgb.R
\name{pick_colour}
\alias{pick_colour}
\title{Pick Colour}
\usage{
pick_colour(variable, alpha = 1, rgb = TRUE)
}
\arguments{
\item{variable}{String selecting the variable of study. One of c('confirmed_cases', 'confirmed_deaths', 'confirmed_recovered')}

\item{alpha}{Alpha value of the final colour}
}
\description{
This function creates a rgba string based on the variable to be studied
}
