% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{plot_all_ts_traces}
\alias{plot_all_ts_traces}
\title{Plots a time series with all traces at once.}
\usage{
plot_all_ts_traces(
  ts_data,
  x_col,
  y_cols,
  y_cols_hidden = c(),
  labels = list()
)
}
\arguments{
\item{ts_data:}{data frame: used when adding different data to an existing plot}

\item{x_col:}{string: name of the x axis}

\item{y_cols:}{character: vector of the col names that will be added to the plot}

\item{labels:}{list: e.g. \code{list(xaxis = "Date", yaxis = "Magnitude", title = "New Plot")}}
}
\value{
plotly plot
}
\description{
This works by building the trace then calling the resulting
function with purrr. Therefore the traces need to have the same x axis.
}
\examples{
\dontrun{
plot_all_ts_traces(
ts_data = data.frame(), 
x = "DATE_REPORTED", 
y = c("NEW_CASES_MA_7DAY", "NEW_CASES_MA_90DAY") 
labels = list(yaxis = "New Cases"))
}
}
