% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh-data.R
\name{refresh_dataset}
\alias{refresh_dataset}
\title{Refreshes a single dataset.}
\usage{
refresh_dataset(dat, output_path)
}
\arguments{
\item{dat}{tibble: df to be written}

\item{output_path}{string: Path to write file to.}
}
\value{
invisible NULL
}
\description{
Refreshes a single dataset.
}
