% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh-data.R
\name{refresh_dataset_from_config}
\alias{refresh_dataset_from_config}
\title{Refreshes a single dataset from a data source config.}
\usage{
refresh_dataset_from_config(data_config)
}
\arguments{
\item{data_config}{A data source config, which is a list containing url and col_types.}
}
\value{
invisible NULL
}
\description{
Refreshes a single dataset from a data source config.
}
