% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file-crud.R
\name{rename_file}
\alias{rename_file}
\title{Rename a file if it exists}
\usage{
rename_file(from, to)
}
\arguments{
\item{from, to}{character vectors, containing file names or paths.
    For \code{file.copy} and \code{file.symlink}
#ifdef windows not active    \code{to} can alternatively be the path to a single existing directory.}
}
\description{
Rename a file if it exists
}
