% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-file-crud.R
\name{safe_write_rds_file}
\alias{safe_write_rds_file}
\title{Write a dataset.}
\usage{
safe_write_rds_file(dat, output_path)
}
\arguments{
\item{dat}{tibble: df to be written}

\item{output_path}{string: Path to write file to}
}
\value{
invisible NULL
}
\description{
Ensure that if function fails original dataset if not lost
}
