% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_table_row.R
\name{side_table_row}
\alias{side_table_row}
\title{Side Table Row}
\usage{
side_table_row(i, top_variables, plus = FALSE)
}
\arguments{
\item{i}{Row of the data.frame to be shown}

\item{top_variables}{A data.frame with the information to be displayed. First column should be the country name and second column the numeric variable.}
}
\description{
This function creates the necessary html to display the tables at the sides of the app
}
