% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dashboard-sidebar.R
\name{sidebar_server}
\alias{sidebar_server}
\title{Dashboard sidebar server function}
\usage{
sidebar_server(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Dashboard sidebar server function
}
