% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-dashboard-sidebar.R
\name{sidebar_ui}
\alias{sidebar_ui}
\title{Dashboard sidebar ui function}
\usage{
sidebar_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Dashboard sidebar ui function
}
