% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-simulations.R
\name{simulations_server}
\alias{simulations_server}
\title{Page: Simulations server function}
\usage{
simulations_server(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Simulations server function
}
