% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-simulations.R
\name{simulations_ui}
\alias{simulations_ui}
\title{Page: Simulations ui function}
\usage{
simulations_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Simulations ui function
}
