% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily-cases-manipulation.R
\name{smooth_daily_agg}
\alias{smooth_daily_agg}
\title{Smooth columns in daily cases data}
\usage{
smooth_daily_agg(daily_agg, cols_used, custom_ma_orders = c())
}
\arguments{
\item{daily_agg}{tibble: daily cases time series (1 date per row)}

\item{custom_ma_orders}{int vector: order of smoothing (each element will produce a new column)}

\item{cols_used:}{character vector: columns to be smoothed}
}
\value{
tibble: the daily cases time series smoothed with original columns
}
\description{
Smooth columns in daily cases data
}
