% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanced_base_utilties.R
\name{source_dir}
\alias{source_dir}
\title{Sources all files in a directory}
\usage{
source_dir(dir, local = FALSE)
}
\arguments{
\item{dir}{string: path of a directory}

\item{local}{boolean or environment: same as \code{source}. TRUE, FALSE or an
environment, determining where the parsed expressions are evaluated. FALSE
(the default) corresponds to the user's workspace (the global environment)
and TRUE to the environment from which source is called.}
}
\value{
NULL (invisible)
}
\description{
Sources all files in a directory
}
\examples{
\dontrun{
source_dir(dir = "./R", local = TRUE)
}
}
\seealso{
\code{source}
}
