% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-time-series-plot.R
\name{time_series_plot_server}
\alias{time_series_plot_server}
\title{Module: Time series plot server function}
\usage{
time_series_plot_server(
  id,
  ts_data,
  labels = list(yaxis = "", box_title = ""),
  deselected_traces = c(),
  ...
)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Module: Time series plot server function
}
