% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-time-series-plot.R
\name{time_series_plot_ui}
\alias{time_series_plot_ui}
\title{Module: Time series plot ui function}
\usage{
time_series_plot_ui(id, show_dropdown = TRUE, collapsible = TRUE, ...)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Module: Time series plot ui function
}
