% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-under-construction.R
\name{under_construction_server}
\alias{under_construction_server}
\title{Under construction placeholder server function}
\usage{
under_construction_server(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Under construction placeholder server function
}
