% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-under-construction.R
\name{under_construction_ui}
\alias{under_construction_ui}
\title{Under construction placeholder ui function}
\usage{
under_construction_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Under construction placeholder ui function
}
