% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-vaccines.R
\name{vaccines_server}
\alias{vaccines_server}
\title{Page: Vaccines server function}
\usage{
vaccines_server(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Vaccines server function
}
