% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-vaccines.R
\name{vaccines_ui}
\alias{vaccines_ui}
\title{Page: Vaccines ui function}
\usage{
vaccines_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Vaccines ui function
}
