% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-world-map.R
\name{world_map_server}
\alias{world_map_server}
\title{Page: World map server function}
\usage{
world_map_server(id, latest_outbreak_rating_by_country)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}

\item{latest_outbreak_rating_by_country}{reactive dataframe}
}
\description{
Page: World map server function
}
