% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-world-map.R
\name{world_map_ui}
\alias{world_map_ui}
\title{Page: World map ui function}
\usage{
world_map_ui(id)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: World map ui function
}
