% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rgb.R
\name{write_rgb}
\alias{write_rgb}
\title{Write RGB}
\usage{
write_rgb(c)
}
\arguments{
\item{c}{Hex code of the colour}
}
\description{
This function converts a hex code into a rgba string, with only the alpha value missing
}
