% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common-docs.R
\docType{data}
\name{common_docs}
\alias{common_docs}
\title{Common documentation}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
common_docs
}
\arguments{
\item{daily_cases}{tibble: daily cases from WHO}

\item{vax}{TODO}

\item{population}{TODO}
}
\description{
Common documentation
}
\keyword{internal}
