% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page-overview.R
\name{overview_server}
\alias{overview_server}
\title{Page: Overview server function}
\usage{
overview_server(id, daily_cases, vax, population)
}
\arguments{
\item{id}{string: indentifier of the function. Used as a namespace.}
}
\description{
Page: Overview server function
}
